added a 64bif fs_offset_t type to clean up most of the mess
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5502 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_demo.c
@@ -189,7 +189,7 @@ void CL_ReadDemoMessage(void)
 			cl.mviewangles[0][i] = LittleFloat(f);
 		}
 
-		if (FS_Read(cls.demofile, net_message.data, net_message.cursize) == (size_t)net_message.cursize)
+		if (FS_Read(cls.demofile, net_message.data, net_message.cursize) == net_message.cursize)
 		{
 			MSG_BeginReading();
 			CL_ParseServerMessage();
